Move MetricFactory to base.py
